9th May

+++ lib/DistanceIR/distanceIR.cpp
@@ -5,7 +5,7 @@ Distanceir :: Distanceir (int pin)
     Analogpin = pin ; 
 }
 
-int Distanceir :: FindDistance ()
+float Distanceir :: FindDistance ()
 {
     int irDistance = analogRead(Analogpin);
     //distance = 0.0001*(irDistance*irDistance) - 0.1439*(irDistance) + 49.9653 ;// replace x as irDistance
